Update user_guide.md
It appears Micro-XRCE-DDS-Agent v2.4.2 has an error in CMakeLists.txt on line 99 where it references fastdds branch 2.12.x which no longer exists. Changing this to 2.12.2 in the fastdds CMakeLists file fixes this, as does using Micro-XRCE-DDS-Agent v2.4.3 which uses 2.14.x which does exist. I was also able to get the sim working on the main branch Micro-XRCE-DDS-Agent, but I'm not sure if there are other consequences of making such a jump with the DDS branch.